+++ src/grape/impl/parsing.clj
@@ -1,0 +1,5 @@
+(ns grape.impl.parsing
+  (:require [parcera.core :as parcera]))
+
+(def parse-code parcera/ast)
+(def unparse-code parcera/code)